url-shortener-flask/requirements.txt
Flask==0.11.1
gunicorn==19.7.0
httplib2==0.9.2
Jinja2==2.8
virtualenv==15.1.0
Werkzeug==0.11.10
requests==2.20.0
mysql-connector==2.2.9       
mysql-connector-python==8.0.20      
mysql-connector-python-rf==2.2.2   
python-dateutil==2.7.5
